Extract BrowserScraper logic to utils
